backend/foodgram_main/requirements.txt
Django
django-filter
djangorestframework
djoser
isort
py
